backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
httpx==0.27.0

# Google Gemini
google-generativeai==0.8.0

# Utils
pydantic[email-validator]==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
aiofiles==24.1.0
